requirements.txt
# Core dependencies
any-agent[langchain]==1.11.0
langchain==0.3.27
openai==2.3.0
jinja2==3.1.6
pydantic==2.12.2
dateparser==1.2.1
PyYAML==6.0.2
langdetect==1.0.9
python-dotenv==1.1.0
fastapi[standard]==0.115.13
langchain-community==0.3.31
langchain-openai==0.3.35
beautifulsoup4==4.13.4
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.25.0
click>=8.0.0
uvicorn>=0.23.1
keybert==0.9.0

# Database dependencies
aiomysql>=0.2.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
redis>=5.0.0
motor>=3.3.0
sqlalchemy[asyncio]>=2.0.0

# Job Queue dependencies
arq>=0.25.0

# Redis resilience dependencies
tenacity>=8.2.0

# Scheduler dependencies
apscheduler>=3.10.0

# Content source dependencies
aiofiles>=23.0.0
aiohttp>=3.8.0
feedparser>=6.0.0
selenium>=4.15.0
watchdog>=3.0.0

# Document processing dependencies
python-docx>=1.2.0
PyPDF2>=3.0.1
